Refactor poller entry code into the poller package
